improved sampling using Quintic Polynomials and modularised code
