+++ frenet+controller+tracker+dummy/src/frenet_planner/config/frenet_planner.yaml
@@ -1,0 +1,25 @@
+frenet_planner:
+    ros__parameters:
+        K_J: 0.1
+        K_T: 0.1
+        K_D: 1.0
+        K_LAT: 1.0
+        K_LON: 1.0
+        MAX_SPEED: 13.89
+        MAX_ACCEL: 5.0
+        MAX_CURVATURE: 1.0
+        MAX_ROAD_WIDTH: 7.0
+        D_ROAD_W: 1.0
+        DT: 0.2
+        MAX_T: 5.0
+        MIN_T: 4.0
+        TARGET_SPEED: 8.33
+        D_T_S: 1.389
+        N_S_SAMPLE: 1.0
+        OBSTACLE_RADIUS: 0.75
+        MIN_LAT_VEL: -5.0 
+        MAX_LAT_VEL: 5.0
+        D_D_NS: 1.0
+        W_X: [0.0 , 10.0 , 20.0 , 30.0 , 40.0 , 50.0 , 60.0 , 70.0 , 80.0]
+        W_Y: [0.0 , 15.0 , 20.0 , 40.0 , 40.0 , 45.0 , 60.0 , 77.0 , 80.0]
+        LOOKAHEAD_DIST: 15.0
