+++ frenet+controller+tracker+dummy/src/frenet_planner/launch/planner.launch.py
@@ -1,0 +1,22 @@
+import os
+from ament_index_python.packages import get_package_share_directory
+from launch import LaunchDescription
+from launch_ros.actions import Node
+
+def generate_launch_description():
+    ld = LaunchDescription()
+    config = os.path.join(
+        get_package_share_directory('frenet_planner'),
+        'config',
+        'frenet_planner.yaml'
+        )
+        
+    node=Node(
+        package = 'frenet_planner',
+        name = 'frenet_planner',
+        executable = 'planner',
+        output="screen",
+        parameters = [config]
+    )
+    ld.add_action(node)
+    return ld
